refactor: move FromRequestParts impl out of api module (#137)
The `axum::extract::FromRequestParts for RequestContext` impl was the only
axum coupling left in the otherwise transport-free `api` module. Move it
into `rest/auth.rs`, next to the `AuthenticationMiddleware` that inserts the
`Principal` it reads back. The `api` module is now fully axum-free; the
`RequestContext` struct and `SecuredAction` stay as domain types.

No behavior change. The remaining items in #137 (a transport-free error path
and relocating the generated axum server glue) are deferred to the actual
crate extraction, where the right shape is forced.
